Update / correct licenseInformation at io-package.json
Adapters which require a paid license in any way need to set this information at io-package.json. This is done by using common.licenseInformation insatead of common.license.

This PR replaces "license":"MIT" which does not match license at README.md and LICENSE files with CC-BY-NC-NO-4.0 and adds the license type "limited" indicating that the full functionality of this adapter requires a paid pro license.

Please check if "license" is OK and change common.licenseInformtion.link to a link where pricing any shop information is visible.